requirements.txt
instagrapi
facebook-scraper
pytesseract
requests
python-dotenv
Pillow
schedule